+++ plugins/postcss-focus-within/test/_browser.mjs
@@ -50,7 +50,7 @@ import { promises as fsp } from 'fs';
 
 	if (!process.env.DEBUG) {
 		const browser = await puppeteer.launch({
-			headless: true,
+			headless: 'new',
 		});
 
 		const page = await browser.newPage();