+++ plugins/postcss-media-minmax/test/_browser.mjs
@@ -74,7 +74,7 @@ import postcss from 'postcss';
 
 	if (!process.env.DEBUG) {
 		const browser = await puppeteer.launch({
-			headless: true,
+			headless: 'new',
 		});
 
 		const page = await browser.newPage();